Remove disable linters in js files
