+++ src/modules/API.js
@@ -1,0 +1,2 @@
+export const baseURL = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api';
+export const gameId = 'ARAId6sN2LKi2AeYJNFw';
